jsBridge代码优化
